mtd: rawnand: toshiba: Pass a single nand_chip object to the status helper
Now that most of the raw NAND API is consistent and and has almost all
its helpers and hooks using a single nand_chip object instead of an
mtd_info one (or both), let's do the same cleanup in the raw NAND
vendors drivers.

Do the same change in the Toshiba driver as well so that the internal
helper to retrieve the ECC status does only take a nand_chip object.

Suggested-by: Boris Brezillon <boris.brezillon@bootlin.com>
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>