of: Update of_device_get_modalias()
This function only needs a "struct device_node" to work, but for
convenience the author (and only user) of this helper did use a "struct
device" and put it in device.c.

Let's convert this helper to take a "struct device node" instead. This
change asks for two additional changes: renaming it "of_modalias()"
to fit the current naming, and moving it outside of device.c which will
be done in a follow-up commit.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>
Reviewed-by: Rob Herring <robh@kernel.org>
---
Hello, please mind that I am on purpose dropping the check for
of_node_reused in the request_module helper because I believe
this check simply does not apply here. If I am misunderstanding
something please let me know.
Thanks!