gpu: host1x: Stop open-coding of_device_uevent()
There is apparently no reasons to open-code of_device_uevent() besides:
- The helper receives a struct device while we want to use the of_node
  from the parent of the struct device.
- of_device_uevent() is not exported to modules.

In practice, the former point is not very constraining, just calling
of_device_uevent(dev->parent...) would have made the trick. The latter
point is more an observation rather than a real blocking point because
nothing prevents of_device_uevent() to be exported to modules, most of
the other helpers in of_device.c actually are, but this one was only
used in drivers/base/ which is built-in anyway.

of_uevent() which is needed by the inline helper of_device_uevent() just
needs to be exported.

Cc: Thierry Reding <thierry.reding@gmail.com>
Cc: David Airlie <airlied@gmail.com>
Cc: Daniel Vetter <daniel@ffwll.ch>
Cc: Rob Herring <robh+dt@kernel.org>
Cc: Frank Rowand <frowand.list@gmail.com>
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>